add EdgeTPUNodeBase for common node base
