+++ python/coral_usb/face_detector.py
@@ -1,5 +1,3 @@
-from dynamic_reconfigure.server import Server
-
 from coral_usb.cfg import EdgeTPUFaceDetectorConfig
 from coral_usb.cfg import EdgeTPUPanoramaFaceDetectorConfig
 from coral_usb.detector_base import EdgeTPUDetectorBase
@@ -7,6 +5,9 @@ from coral_usb.detector_base import EdgeTPUPanoramaDetectorBase
 
 
 class EdgeTPUFaceDetector(EdgeTPUDetectorBase):
+
+    _config_class = EdgeTPUFaceDetectorConfig
+
     def __init__(self, namespace='~'):
         super(EdgeTPUFaceDetector, self).__init__(None, None, namespace)
 
@@ -14,17 +15,11 @@ class EdgeTPUFaceDetector(EdgeTPUDetectorBase):
         self.label_ids = [0]
         self.label_names = ['face']
 
-    def start_dynamic_reconfigure(self, namespace):
-        # dynamic reconfigure
-        dyn_namespace = namespace
-        if namespace == '~':
-            dyn_namespace = ''
-        self.srv = Server(
-            EdgeTPUFaceDetectorConfig,
-            self.config_callback, namespace=dyn_namespace)
-
 
 class EdgeTPUPanoramaFaceDetector(EdgeTPUPanoramaDetectorBase):
+
+    _config_class = EdgeTPUPanoramaFaceDetectorConfig
+
     def __init__(self, namespace='~'):
         super(EdgeTPUPanoramaFaceDetector, self).__init__(
             None, None, namespace)
@@ -32,12 +27,3 @@ class EdgeTPUPanoramaFaceDetector(EdgeTPUPanoramaDetectorBase):
         # only for human face
         self.label_ids = [0]
         self.label_names = ['face']
-
-    def start_dynamic_reconfigure(self, namespace):
-        # dynamic reconfigure
-        dyn_namespace = namespace
-        if namespace == '~':
-            dyn_namespace = ''
-        self.srv = Server(
-            EdgeTPUPanoramaFaceDetectorConfig,
-            self.config_callback, namespace=dyn_namespace)
